# wikify-web/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wikify Web API",
    "description": "AI-powered repository documentation and chat system",
    "contact": {
      "name": "Wikify Team",
      "email": "support@wikify.dev"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.wikify.dev",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/repositories": {
      "get": {
        "tags": [
          "Repository"
        ],
        "summary": "List repositories",
        "description": "List all repositories accessible to the current user",
        "operationId": "list_repositories",
        "responses": {
          "200": {
            "description": "Repositories listed successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Repository"
        ],
        "summary": "Initialize repository",
        "description": "Initialize a repository for processing and create a new session. If the repository is already being indexed by another session, an error will be returned.",
        "operationId": "initialize_repository",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitializeRepositoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Repository initialized successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitializeRepositoryResponse"
                }
              }
            }
          },
          "409": {
            "description": "Repository is already being indexed by another session"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/repositories/{repository_id}": {
      "get": {
        "tags": [
          "Repository"
        ],
        "summary": "Get repository information",
        "description": "Get information about a repository",
        "operationId": "get_repository_info",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "description": "Repository ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repository information retrieved successfully"
          },
          "404": {
            "description": "Repository not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Repository"
        ],
        "summary": "Delete repository",
        "description": "Delete a repository and all associated data including sessions, vector data, and database records",
        "operationId": "delete_repository",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "description": "Repository ID to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repository deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteRepositoryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Repository not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/repositories/{repository_id}/reindex": {
      "post": {
        "tags": [
          "Repository"
        ],
        "summary": "Reindex repository",
        "description": "Reindex an existing repository. If the repository is currently being indexed, returns a conflict error. If already indexed, resets the state and starts reindexing.",
        "operationId": "reindex_repository",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "description": "Repository ID to reindex",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repository reindexing started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitializeRepositoryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Repository not found"
          },
          "409": {
            "description": "Repository is currently being indexed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteRepositoryResponse": {
        "type": "object",
        "description": "Repository deletion response",
        "required": [
          "status",
          "message",
          "deleted_repository_id"
        ],
        "properties": {
          "deleted_repository_id": {
            "type": "string",
            "example": "uuid-string"
          },
          "message": {
            "type": "string",
            "example": "Repository deleted successfully"
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
...
      "InitializeRepositoryRequest": {
        "type": "object",
        "description": "Repository initialization request",
        "required": [
          "repository"
        ],
        "properties": {
          "access_token": {
            "type": [
              "string",
              "null"
            ]
          },
          "auto_generate_wiki": {
            "type": [
              "boolean",
              "null"
            ],
            "example": true
          },
          "auto_index": {
            "type": [
              "boolean",
              "null"
            ],
            "example": true
          },
          "metadata": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "repo_type": {
            "type": [
              "string",
              "null"
            ],
            "example": "github"
          },
          "repository": {
            "type": "string",
            "example": "https://github.com/user/repo"
          }
        }
      },
      "InitializeRepositoryResponse": {
        "type": "object",
        "description": "Repository initialization response",
        "required": [
          "repository_id",
          "status",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Repository initialized successfully"
          },
          "repository_id": {
            "type": "string",
            "example": "uuid-string"
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}